feat: implement BillsHistoryPage component and integrate it into the dashboard
