+++ src/app/dashboard/page.tsx
@@ -8,7 +8,7 @@ export default async function Page() {
 	const loggedInUser = await getUser();
 	const availableProviders = await getUtilityProviders(loggedInUser.id);
 	// const currentDate = new Date();
-	const currentDate = new Date("2025-06-01");
+	const currentDate = new Date("2025-05-01");
 	const fetchedBills = await fetchUserBills(
 		availableProviders,
 		currentDate.getMonth() + 1, // getMonth() is zero-based
